[#146] rename: enum들을 모두 constant 폴더로이동
