+++ collector/src/main/java/com/wherecar/collector/application/CarLogConverterServiceImpl.java
@@ -3,10 +3,10 @@ package com.wherecar.collector.application;
 import com.wherecar.collector.domain.Car;
 import com.wherecar.collector.domain.CarLog;
 import com.wherecar.collector.domain.CarStatus;
-import com.wherecar.collector.domain.GpsConditionType;
+import com.wherecar.collector.domain.constant.GpsConditionType;
 import com.wherecar.collector.application.dto.CarLogRequest;
 import com.wherecar.collector.application.dto.CarLogResponse;
-import com.wherecar.collector.application.dto.ResponseCode;
+import com.wherecar.collector.application.dto.constant.ResponseCode;
 import com.wherecar.collector.infrastructure.CarLogRepository;
 import com.wherecar.collector.infrastructure.CarRepository;
 import com.wherecar.collector.infrastructure.CarStatusRepository;
