+++ collector/src/main/java/com/wherecar/collector/application/dto/constant/ResponseCode.java
@@ -1,4 +1,4 @@
-package com.wherecar.collector.application.dto;
+package com.wherecar.collector.application.dto.constant;
 
 import lombok.Getter;
