+++ collector/src/main/java/com/wherecar/collector/domain/CarLog.java
@@ -1,5 +1,7 @@
 package com.wherecar.collector.domain;
 
+import com.wherecar.collector.domain.constant.DriveType;
+import com.wherecar.collector.domain.constant.GpsConditionType;
 import jakarta.persistence.*;
 import lombok.*;
