+++ collector/src/main/java/com/wherecar/collector/domain/constant/AcquisitionType.java
@@ -1,4 +1,4 @@
-package com.wherecar.collector.domain;
+package com.wherecar.collector.domain.constant;
 
 public enum AcquisitionType {
     PURCHASE,
